refactor: ページのアーカイブ機能を削除し、削除済みページの管理を実装 (#1083)

+++ src/app/[locale]/(common-layout)/_components/page/delete-page-dialog/action.ts
@@ -1,11 +1,10 @@
 "use server";
 import { z } from "zod";
 import { deleteActionFactory } from "@/app/[locale]/_action/delete-action-factory";
-import { archivePage } from "./db/mutations.server";
-export const archivePageAction = deleteActionFactory({
+import { deletePage } from "./db/mutations.server";
+export const deletePageAction = deleteActionFactory({
 	inputSchema: z.object({
 		pageId: z.coerce.number(),
 	}),
-	deleteById: ({ pageId }, userId) =>
-		archivePage(pageId, userId).then(() => {}),
+	deleteById: ({ pageId }, userId) => deletePage(pageId, userId).then(() => {}),
 });
